Componentise App Indicators section
#2607
