+++ waltz-ng/client/applications/app-view.html
@@ -71,8 +71,7 @@
 
 
     <!-- INDICATORS -->
-    <waltz-entity-statistic-section name="Indicators"
-                                    entity-statistics="ctrl.entityStatistics">
+    <waltz-entity-statistic-section parent-entity-ref="ctrl.entityRef">
     </waltz-entity-statistic-section>
